Incorporate the autofree system from nica/sol/cve/etc
Where possible we assign autofree (attribute-based cleanups) for our own types,
to ensure that cleanup and freeing of memory always happens as soon as they
go out of scope. This allows a much cleaner codebase with a fail-early
approach, and avoids the spaghetti branch cleaning often seen in older C
projects.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>