cli/configure: Implement the NVIDIA vs NVIDIA Optimus paths
When we encounter both intel & nvidia, we check if the intel graphics is
the one used by the BIOS during boot. In this case we determine we're
dealing with Optimus, and appropriately proxy back off to the optimus
specific GPU function.

In the instance we have NVIDIA and Intel graphics, but the NVIDIA is the
boot_vga, we determine that it is a non-optimus "simple" GPU configuration.
This may in fact still be Optimus, but the BIOS has switched the GPU over
as an option. We treat this as a simple case and only configure the NVIDIA
driver.

The next steps would be to detect AMD hybrid, with similar logic, and then
fallback to the boot_vga device from a complex situation to reduce back
to the primary GPU and configure drivers for that only.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>