Identify USB interfaces, don't skip modalias-less devices
This allows us to identify the parent device which has no modalias,
such as a usb_device, before we identify the child interfaces. This will
also allow us to then add interfaces to a parent device in future.

Additionally we stop using hwdb directly with a modalias lookup, and
instead grab the properties straight from the device itself.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>