plugin: Implement main LdmModaliasPlugin detection routine
This one is very trivial to implement, effectively walk our modaliases
until we find one that actually matches, and return a new LdmProvider
at that point.

We'll need to flesh out the API of Provider to have package name and
kernel module, etc, but this is the bulk of that work now done.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>