Add support for HID detection
We add a staging type, LdmHIDDevice, which allows us to perform some special
adaptation for HID support. Currently we don't flesh this out too well and
in reality we could extend this in future to differentiate between a keyboard
and mouse, etc.

Right now we're only interested in USB devices that are extended through
an interfaces HID child, which updates the entire device definition to
include the HID child. Additionally we now support recursive child matching
so that we're able to identify `hid:` style modaliases. To verify this, we
add a new USB specific test case for the Razer driver, and prove we're able
to not only discover the root level USB device, but find the corresponding
razer-drivers package as a modalias match.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>