Rename `mkmodalias` to `mkmodaliases`
Pluralisation sucks, but there is a difference to be considered here.
The kernel knows a modalias, a device and a module know a modalias, and
a module may support one or more modaliases. Our files contain modalias
definitions for pattern matching, thus they are modaliases, plural, to
match against singular modalias entries on a device.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>