Add skeleton for the new LdmGLXManager class
Instead of repeating history with the CLI command we're instead going
to convert the code into the library portion for managing GLX providers,
keeping the CLI clean and simple. This also means that the LDM system
can be extended upon by others trivially without them being bound to
the CLI tools we provide (i.e. for other distros)

We'll use this for handling the usual logic of symlink mangling and
dynamically doing the "right thing" when it comes to hooking up X.Org,
etc.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>