Add support for writing "simple" GPU configuration xorg configs
This writes a static snippet out to the named path which is used for
trivial X11 configuration files, i.e. telling a modern X server which
driver to use.

This isn't necesarily required on all distros, as some will patch their
X server to automatically load `nvidia` or such if it exists. Either way,
this is the file that we manage and will continue to write, as it's also
the one we'll write when setting up Optimus in future.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>