Add initial helper wrappers to simplify Optimus logic
We're going to choose here to introduce the LDM Early Init into all
display managers, regardless of whether or not they're using Optimus.
This will allow us to make a very lightweight entry point that can
decide whether or not to execute the appropriate xrandr helpers based
on whether hybrid is enabled in some control file (not yet done).

Signed-off-by: Ikey Doherty <ikey@solus-project.com>