Add new session-init binary (part 2)
This binary will only undergo the full motions if the hybrid tracking
file is present, and will simply modify xrandr within the session to
match the requirements for Optimus devices.

Right now this is completely dependent on NVIDIA Optimus proprietary
drivers, but could be extended for other configurations if needed.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>