Remove in-tree implementation of hardware monitor
As LDM itself is not "allowed" to know what a package is, by design, it
should therefore not have an implemented hotplug daemon. We can perhaps
provide some reference material in terms of examples, but we won't directly
implement one in tree.

Instead it should be implemented in the software center or some background
service that knows what a package is.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>