Entirely banish Gio from the codebase
It makes stuff leak by default, and initialises gdbus, etc, which is entirely
inappropriate for backend system like the driver configuration in the CLI.
Instead we use POSIX APIs in place of the explicitly "gio" portions and
kick it out of the codebase.

Unfortunately this means we break ABI by nuking the symbol, however we've
not yet stabilised it, so, yay?

Signed-off-by: Ikey Doherty <ikey@solus-project.com>