lib: Pass the LdmDevice to the signal before we remove
It feels hella inconsistent to pass a string in the removal signal when
we pass a device in the add signal. So, break the ABI again and instead
we'll emit the device during removal which makes it easier to work with.
After this is done we'll still unref it, but we ensure that it is valid
during our signal context.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>