manager: Make the get_devices API more resilient/consistent
In keeping with the primary APIs, i.e. getting providers, we switch to
a GPtrArray for our get_devices API. This allows us to explicitly set
our own free function and re-ref each member of the returned store. This
should help with race conditions where a device is known to a client but
the hotplug code removes it. In this case, the last refcount instance
will be in the domain of the API consumer, and the library will have
already "forgotten" about it.

This doesn't introduce a leak, however, as we ensure to remove all child
node devices when removing the toplevel device.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>