lib: Replace signal child logic with recursive has_ functions
The old logic was tied too deeply to the USB implementation, so we now
make the functions recursive into the child tree. This means we can get
rid of USB "original class" specifics and probe a tree from the toplevel
node and still get the same results. Additionally this will allow us to
support PCI device children as we go forward.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>